Advanced statistics (HLL, MCV, unified rebuild) and FieldEvaluator optimization (#144)
Add HyperLogLog, MostCommonValues, and StatisticsRebuilder for selectivity
estimation. Replace BasicSelectivityEstimator with AdvancedSelectivityEstimator
(MCV → Histogram → B+Tree → uniform chain). Add StatisticsWorker background
thread, mutation counter, and NavigationView/OrView query support.

+++ test/Typhon.Engine.Tests/Data/Query/ViewTests.cs
@@ -15,8 +15,8 @@ class ViewTests : TestBase<ViewTests>
         KeyType keyType, CompareOp compareOp, long threshold) =>
         new()
         {
-            FieldIndex = fieldIndex,
-            FieldOffset = fieldOffset,
+            FieldIndex = (byte)fieldIndex,
+            FieldOffset = (ushort)fieldOffset,
             FieldSize = fieldSize,
             KeyType = keyType,
             CompareOp = compareOp,
